screen event classes collected
